Provide default MIME type and pathTranslation
Provide default MIME type rather than not upload files.  text/plain
seems like a good default as it fits all the files we use that
currently are not properly detected.

Provide a default pathTranslation that does nothing.

Make pathToIndex optional, not returning externalUrl if it is not
provided.

If no AWS credentials are in the SDM configuration, fall back to the
AWS default credentials chain.

Closes #3